🐛 Fix tree style tabs hidden behind browser page
Fix the z-index of the tree style tab sidebar. Before, it was a static
value of 1, which put it in front of the browser page. A recent change
to Firefox changed the z-index of the browser page, resulting in the
sidebar being hidden behind the page when expanded.

This was fixed by computing the z-index of the sidebar from the
css variable used for the browser page's z-index.